requirements.txt
Django>=4.1.5, <4.2
djangorestframework>=3.14.0 ,<3.15
drf-spectacular>=0.25.1,<0.26
psycopg2>=2.9.5, <3.0.0
django-phonenumber-field >=7.0.2, <7.1
phonenumbers >= 8.13.5, <8.14
